chore: move reorder perf test to separate benchmark module
The 800-page/80-group perf test was timing out CI with too many writes.
Moved to wiki/benchmarks/ so it only runs on-demand.
